Moving consumer file to app
